requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cx-Freeze==6.1
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.2.0
django-guardian==2.2.0
django-phone-field==1.8.0
django-post-office==3.4.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.1
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==4.6.10
Markdown==3.2.2
MarkupSafe==1.1.1
more-itertools==8.2.0
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
somepackage==1.2.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.23
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
whitenoise==5.0.1
zipp==3.1.0
